Add test cases to test_create_st
